Add ApplicationManager
