+++ addressbook-web-tests/src/test/java/ru/ls/qa/school/addressbook/NewContactTest.java
@@ -2,22 +2,19 @@ package ru.ls.qa.school.addressbook;// Generated by Selenium IDE
 
 import org.junit.jupiter.api.Test;
 
-import static com.codeborne.selenide.Selenide.$;
-import static com.codeborne.selenide.Selenide.open;
-
 public class NewContactTest extends TestBase {
 
     @Test
     public void testNewContact() {
-        goToNewContactPage();
-        fillNewContactForm(new ContactData("testname",
+        app.goToNewContactPage();
+        app.fillNewContactForm(new ContactData("testname",
                 "testname",
                 "testname",
                 "testname",
                 "Moscow"));
-        submitCreationNewContact();
-        returnToMainPage();
-        logout();
+        app.submitCreationNewContact();
+        app.returnToMainPage();
+        app.logout();
     }
 }
